Move FR doc endpoint to V3 and reorganize views, serializers
