+++ solution/backend/resources/utils.py
@@ -1,11 +1,3 @@
-def is_int(x):
-    try:
-        _ = int(x)
-        return True
-    except ValueError:
-        return False
-
-
 # TODO: remove this after v3 move
 class reverse_sort:
     def __init__(self, obj):
